recompiler_new: signext fix (attempt)
svn:sync-xref-src-repo-rev: r40317

+++ src/recompiler_new/VBoxRecompiler.c
@@ -1,4 +1,4 @@
-/* $Id: VBoxRecompiler.c 14755 2008-11-28 02:58:01Z knut.osmundsen@oracle.com $ */
+/* $Id: VBoxRecompiler.c 14956 2008-12-03 20:34:30Z knut.osmundsen@oracle.com $ */
 /** @file
  * VBox Recompiler - QEMU.
  */
@@ -3113,7 +3113,7 @@ void remR3PhysRead(RTGCPHYS SrcGCPhys, void *pvDst, unsigned cb)
  *
  * @param   SrcGCPhys       The source address (guest physical).
  */
-uint8_t remR3PhysReadU8(RTGCPHYS SrcGCPhys)
+uint32_t remR3PhysReadU8(RTGCPHYS SrcGCPhys)
 {
     uint8_t val;
     STAM_PROFILE_ADV_START(&gStatMemRead, a);
@@ -3132,7 +3132,7 @@ uint8_t remR3PhysReadU8(RTGCPHYS SrcGCPhys)
  *
  * @param   SrcGCPhys       The source address (guest physical).
  */
-int8_t remR3PhysReadS8(RTGCPHYS SrcGCPhys)
+int32_t remR3PhysReadS8(RTGCPHYS SrcGCPhys)
 {
     int8_t val;
     STAM_PROFILE_ADV_START(&gStatMemRead, a);
@@ -3151,7 +3151,7 @@ int8_t remR3PhysReadS8(RTGCPHYS SrcGCPhys)
  *
  * @param   SrcGCPhys       The source address (guest physical).
  */
-uint16_t remR3PhysReadU16(RTGCPHYS SrcGCPhys)
+uint32_t remR3PhysReadU16(RTGCPHYS SrcGCPhys)
 {
     uint16_t val;
     STAM_PROFILE_ADV_START(&gStatMemRead, a);
@@ -3170,7 +3170,7 @@ uint16_t remR3PhysReadU16(RTGCPHYS SrcGCPhys)
  *
  * @param   SrcGCPhys       The source address (guest physical).
  */
-int16_t remR3PhysReadS16(RTGCPHYS SrcGCPhys)
+int32_t remR3PhysReadS16(RTGCPHYS SrcGCPhys)
 {
     uint16_t val;
     STAM_PROFILE_ADV_START(&gStatMemRead, a);
